+++ crates/cosmos/cosmos-integration-tests/src/impls/bootstrap/build_cosmos_chain_driver.rs
@@ -83,7 +83,7 @@ where
             .clone();
 
         let chain = bootstrap
-            .build_chain_with_node_config(&chain_node_config, &relayer_wallet)
+            .build_chain_with_node_config(&chain_node_config, &genesis_config, &relayer_wallet)
             .await?;
 
         let chain_command_path = bootstrap.chain_command_path().clone();